worker/requirements.txt
flask==2.3.3
numpy==1.24.3
pillow==10.0.0
numba==0.57.1
requests==2.31.0
gunicorn==21.2.0